+++ packages/components/src/icon/stories/index.story.tsx
@@ -20,7 +20,7 @@ const meta: ComponentMeta< typeof Icon > = {
 	component: Icon,
 	parameters: {
 		controls: { expanded: true },
-		docs: { source: { state: 'open' } },
+		docs: { canvas: { sourceState: 'shown' } },
 	},
 };
 export default meta;